fix: gate tunnel forward on admission and split safety-floor tests
Tunnel ServeHTTP now refuses work-producing Forward when admission is
closed, matching the API mutation gate. Split the 698-line safety_floor
test file into admission/scheduler/recovery files under the 300-line cap.

Generated-By: looper 0.11.0-beta.1 (runner=fixer, agent=grok-build)
