+++ internal/runtime/runtime.go
@@ -272,6 +272,9 @@ func New(options Options) *Runtime {
 	}
 	if rt.webhook != nil {
 		rt.webhook.forwarder = rt.WebhookForwarder
+		// Tunnel listener is not behind the API mutation gate; consult the same
+		// admission Authority before enqueueing reviewer/fixer work (#583).
+		rt.webhook.allowForward = rt.AllowMutations
 	}
 	if !customSchedulerTick {
 		rt.runSchedulerTick = rt.executeDefaultSchedulerTick
